% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{generate_map_pc_chord}
\alias{generate_map_pc_chord}
\title{Generate pc_chord map}
\usage{
generate_map_pc_chord()
}
\description{
This function generates mappings between pc_chord IDs and derived features.
}
