% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{generate_transpose_pc_chord_id}
\alias{generate_transpose_pc_chord_id}
\title{Generate transposition mapping}
\usage{
generate_transpose_pc_chord_id()
}
\description{
Generates a mapping between pitch-class chords and their transpositions.
For more information see \code{\link{pc_chord_transpositions}}.
}
