% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvrmap-package.R
\docType{package}
\name{hvrmap-package}
\alias{hvrmap}
\alias{hvrmap-package}
\title{hvrmap: Defining Maps for the 'hvr' Package}
\description{
Supports the 'hvr' package by defining mappings between 'pc_chord' objects and derived features.
}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
