% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvrmap-package.R
\docType{data}
\name{pc_chord_transpositions}
\alias{pc_chord_transpositions}
\title{Pitch-class chord transpositions}
\description{
This 12 x 24,576 matrix constitutes a look-up table for efficiently
transposing vectors of pitch-class chords,
which must be encoded using the integer mapping scheme
from the \code{hrep} package.
The element in the ith row and the jth column corresponds to the
(encoded) pitch-class chord produced by transposing
pitch-class chord j by pitch-class interval i.
}
\keyword{data}
