% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose_pc_chord_id}
\alias{transpose_pc_chord_id}
\title{Transpose pitch-class chord}
\usage{
transpose_pc_chord_id(pc_chord_id, interval)
}
\arguments{
\item{pc_chord_id}{An integer vector corresponding to pitch-class chords as encoded
using the \code{hrep} package (https://pmcharrison.github.io/hrep).}

\item{interval}{A single integer between 0 and 12, corresponding to the pitch-class
interval by which the pitch-class chords ought to be transposed.}
}
\value{
An integer vector corresponding to transposed pitch-class chords
encoded using the encoding scheme from the \code{hrep} package.
}
\description{
Transposes a series of pitch-class chords by a given interval.
}
